Support for forgetting AsyncResults to redis Redis memory consumption.
This required redesigning report/email generation since we will no longer be able to reliably walk all results. We now store a list of tasks in the db for which a report (e.g. @email) is expected. Nodes that have reports will not be deleted by default, and that db-stored list is used to retrieve results of these nodes at time of report generation.

Trees rooted at nodes that executed with enqueue_once are similarly not forgotten by default.